разработал экран авторизованного пользователя

+++ app/src/main/java/ru/example/zencartest/db/dao/UserDao.kt
@@ -19,5 +19,5 @@ interface UserDao {
     suspend fun insert(userEntity: UserEntity)
 
     @Query("DELETE FROM users WHERE id = :id")
-    suspend fun removeById(id: Long)
+    suspend fun removeById(id: Int)
 }
